Revert "fix"
This reverts commit bed78613b8bb0f8fa28d8df549d30971f59222e6.

+++ core/interaction-api/src/main/java/ru/yandex/practicum/dto/event/UpdateEventAdminDto.java
@@ -26,7 +26,7 @@ public class UpdateEventAdminDto {
     private LocationDto location;
     private Boolean paid;
     @PositiveOrZero
-    private Long participantLimit;
+    private Integer participantLimit;
     private Boolean requestModeration;
     private AdminEventAction stateAction;
     @Size(min = 3, max = 120)
